+++ chess_engine/src/coordinates.cpp
@@ -7,8 +7,8 @@ std::optional<Coordinates> Coordinates::from_str(std::string_view str)
     return {};
   }
 
-  auto x = static_cast<int8_t>(toupper(str[0]) - 'A');
-  auto y = static_cast<int8_t>(str[1] - '1');
+  auto x = static_cast<int32_t>(toupper(str[0]) - 'A');
+  auto y = static_cast<int32_t>(str[1] - '1');
   if (x < 0 || x > 7 || y < 0 || y > 7)
   {
     return {};
